Rework metrics settings to merge similar metrics
This greatly simplifies the implementation, and improves UX by removing
the weird coupling between bench settings and metrics settings.

A metric can now be configured even if it is not provided by any
currently configured instrument. If an instrument is added which
provides matching metrics, they will be automatically tracked.