Rework metrics settings to merge similar metrics (#75)
* Rework metrics settings to merge similar metrics

This greatly simplifies the implementation, and improves UX by removing
the weird coupling between bench settings and metrics settings.

A metric can now be configured even if it is not provided by any
currently configured instrument. If an instrument is added which
provides matching metrics, they will be automatically tracked.

* Re-implement Metric TypeDataProvider

It seems this is required by the runner because it needs to make a
name-based type lookup of a given metric.

* Finish renaming descriptors/discriminants to specifiers

* Fix bug causing pollrate to reset when deserializing

* only support GetTypeData from string

* Fallback metric name

* Implement metric typedata source

* Add validation errors for CurrentSources


* Update CurrentSources implementation

Update logic for getting the name of a metric

* Remove blocklist and add enabled flag to metric settings

* Make Instrument base classes implement IMetricSource to make sure their metrics gets discovered

* Nats poller: filter distinct metrics

Currently, there is a validation error if a duplicate metric is added,
but this is not enforced. The poller should ensure it does not poll
duplicate metrics.

---------

Co-authored-by: Asger Iversen <20689829+AsgerIversen@users.noreply.github.com>