add example implementation of metric sink
This implementation contains two variants of metric sinks:
1. A MetricSinkSettings system where sinks are configured separately.
This should be a simplification for e.g. the Nats sink because we can
configure the nats sink in the default session, and just leave it
disabled in other sessions. Then there should be no need for the UI to
activate it. This also gives sinks the opportunity to provide config
options.
2. A system that automatically instantiates all IMetricSinks at startup.
This is maybe a bit simpler (all sinks are always available), but is a
lot less flexible.