avoid serializing merged metric information
Merged metric information is information about a metric specifier which
must be aligned between different metric instances. This data must not
be serialized because it depends on the constellation of installed
plugins. For example, if a plugin provides CalibrationDate as a Push
metric, this could promote CalDate to a PushPoll metric. But if this
plugin is later uninstalled, it should be demoted again. This also
applies to default metrics / default poll rate of metrics.